requirements.txt
pygame==2.4.0
pyserial==3.5
